+++ app/src/components/MapCard/index.js
@@ -1,6 +1,5 @@
 import React, {Component} from 'react';
-import {Animated, Dimensions, Image, StyleSheet, Text, TouchableOpacity, ImageBackground, View} from 'react-native';
-import {Button, Icon} from 'native-base';
+import {Animated, Dimensions, Image, ImageBackground, StyleSheet, Text, TouchableOpacity, View} from 'react-native';
 import Interactable from 'react-native-interactable';
 import LinearGradient from 'react-native-linear-gradient';
 import StreetView from 'react-native-streetview';
@@ -25,10 +24,10 @@ const Screen = {
 };
 
 export default class MapCard extends Component {
-  renderIcon = (state, iconName, iconFamily, text) => {
+  renderIcon = (state, icon, text) => {
     return state ? (
       <View style={styles.iconContainer}>
-        <Icon type={iconFamily} name={iconName} style={styles.icon}/>
+        <Image style={styles.icon} source={icon}/>
         <Text style={styles.iconText}>{text}</Text>
       </View>
     ) : null;
@@ -51,6 +50,8 @@ export default class MapCard extends Component {
     }
 
     const distance = humanizeDistance(this.props.parkspot.dist);
+
+    const anyIconShown = this.props.parkspot.electricCharger || this.props.parkspot.accessible;
 
     return (
       <View style={styles.panelContainer} pointerEvents={'box-none'}>
@@ -89,21 +90,23 @@ export default class MapCard extends Component {
               style={styles.panelSubtitle}>{this.props.parkspot.street} {this.props.parkspot.houseNumber}, {this.props.parkspot.city}, {this.props.parkspot.country}</Text>
             <LinearGradient style={styles.panelGradient} colors={gradient.colors} start={gradient.start}
                             end={gradient.end} locations={gradient.locations}>
-              <TouchableOpacity block style={[styles.panelButton, styles.buttonShadow]} onPress={this.props.onStartNavigation}>
-                <ImageBackground style={styles.panelImage} source={require('../../../assets/navigationButton.png')}>
-                    <Text style={styles.panelButtonTitle}>Go there!</Text>
+              <TouchableOpacity block style={[styles.panelButton, styles.buttonShadow]}
+                                onPress={this.props.onStartNavigation}>
+                <ImageBackground style={styles.panelImage} source={require('../../../assets/buttons/navigation.png')}>
+                  <Text style={styles.panelButtonTitle}>Go there!</Text>
                 </ImageBackground>
               </TouchableOpacity>
             </LinearGradient>
+            (state, icon, text)
 
             <View style={styles.iconsContainer}>
-              {this.renderIcon(this.props.parkspot.electricCharger, 'ios-flash', 'Ionicons', 'Charging possible')}
-              {this.renderIcon(this.props.parkspot.accessible, 'accessibility', 'MaterialIcons', 'Easily accessible')}
+              {this.renderIcon(this.props.parkspot.electricCharger, require('../../../assets/icons/filter/electricCharger.png'), 'Charging possible')}
+              {this.renderIcon(this.props.parkspot.accessible, require('../../../assets/icons/filter/accessible.png'), 'Easily accessible')}
             </View>
 
             <View style={styles.streetViewContainer}>
               <StreetView
-                style={styles.streetView}
+                style={[styles.streetView, {marginTop: anyIconShown ? 12 : 54}]}
                 allGesturesEnabled={true}
                 coordinate={{
                   'latitude': parseFloat(this.props.parkspot.lat),